refactor(synapsys): extract cite-scan helpers to separate module
Extracted 8 cite-scan helper functions (readFiredMemoryNames,
transcriptRowToText, extractFromTranscript, extractResponseText,
parseCiteSignalsList, recoverCiteSignals, emitCitations, runCiteScan)
from hooks/synapsys.js into new lib/cite-scan.js. Dispatcher now ~176
lines shorter, below 400-line quality cap breached by prior main merge.
No behavior change; all fail-open semantics preserved. Exports all
helpers for unit testing.
